Upgrade data to a framework
